moved unique id method
